refactor(ui): rename SourceDrawer file to be PascalCase too
